+++ src/TestExecutableInfo.ts
@@ -10,6 +10,11 @@ import { TestSuiteInfoFactory } from './TestSuiteInfoFactory';
 import { SharedVariables } from './SharedVariables';
 import { GazeWrapper, VSCFSWatcherWrapper, FSWatcher } from './FSWatcher';
 
+export interface TestExecutableInfoFrameworkSpecific {
+  helpRegex?: string;
+  additionalRunArguments?: string[];
+}
+
 export class TestExecutableInfo implements vscode.Disposable {
   public constructor(
     private readonly _shared: SharedVariables,
@@ -23,6 +28,9 @@ export class TestExecutableInfo implements vscode.Disposable {
     private readonly _defaultCwd: string,
     private readonly _defaultEnv: { [prop: string]: string },
     private readonly _variableToValue: ResolveRulePair[],
+    private readonly _catch2: TestExecutableInfoFrameworkSpecific,
+    private readonly _gtest: TestExecutableInfoFrameworkSpecific,
+    private readonly _doctest: TestExecutableInfoFrameworkSpecific,
   ) {
     this._name = name !== undefined ? name : '${filename}';
     this._description = description !== undefined ? description : '${relDirpath}/';
@@ -44,7 +52,7 @@ export class TestExecutableInfo implements vscode.Disposable {
   public async load(): Promise<void> {
     const pattern = this._patternProcessor(this._pattern);
 
-    this._shared.log.info('pattern', pattern);
+    this._shared.log.info('pattern', this._pattern, this._shared.workspaceFolder.uri.fsPath, pattern);
 
     if (pattern.isAbsolute && pattern.isPartOfWs)
       this._shared.log.info('Absolute path is used for workspace directory. This is unnecessary, but it should work.');
@@ -284,10 +292,19 @@ export class TestExecutableInfo implements vscode.Disposable {
       this._shared.log.error('resolvedEnv', e);
     }
 
-    return new TestSuiteInfoFactory(this._shared, resolvedName, resolvedDescription, filePath, {
-      cwd: resolvedCwd,
-      env: Object.assign({}, process.env, resolvedEnv),
-    }).create();
+    return new TestSuiteInfoFactory(
+      this._shared,
+      resolvedName,
+      resolvedDescription,
+      filePath,
+      {
+        cwd: resolvedCwd,
+        env: Object.assign({}, process.env, resolvedEnv),
+      },
+      this._catch2,
+      this._gtest,
+      this._doctest,
+    ).create();
   }
 
   private _handleEverything(filePath: string): void {